build: Externalize versioning into a file
This patch moves the project version from being hardcoded in
CMakeLists.txt to a separate VERSION file (and set the version to 0.99.0).
This makes version management easier, especially for automated scripts
that need to read or update the version number.

Also add a new CMake function, parse_version_file, is introduced
to read the KEY = VALUE pairs from the VERSION file.

The project() call is updated to use the version information
parsed from this new file.

The minimum required CMake version has been bumped to 3.20.

Signed-off-by: Andrey VOLKOV <andrey@volkov.fr>